Include amount in test md5
There is currently an issue with the hash that is related to the amount.
Because tests didn't include the amount in their hash values, this didn't manifest in a test failure.
The tests are now using the default md5 values, which includes amount.
Probably even better would be to use ALL possible fields in the tests md5, even if not default.
For now, this should catch the most important bugs though.