Revert change in number format
A previous commit changed the format of 6.5 to 6.50.
While this was intentional, it also broke the md5 hash duplication detection, meaning it would not detect duplicates created by earlier versions of fints2ledger.

While I would like to change the format to always included 2 decimal places, it doesn't not seem worth it if this is a backwards incompatible change because of the duplication detection.