Deactivate static building for now
Because it fails and I can't find a good solution for it.
Restructuring the workflow more would probably lead to a solution (like
using GHCup instead of the haskell setup actions, which allows running
on alpine like hledger does), but maybe not statically linking is also
perfectly fine. Pandoc for example doesn't statically link.